pyproject.toml
[project]
name = "unlinked"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.52.0",
    "pydantic>=2.12.5",
    "pynput>=1.8.1",
    "requests>=2.32.5",
]
